// src/enums/status.enum.ts
export enum StatusEnum {
    CREATED = 1,
    PENDING = 3,
    IN_PROCCESS = 5,
    SEND_TO_CONFIRM = 7,
    ROLLED_BACK = 9,
    REJECTED = 11,
    REQUEST_EXTENSTION = 13,
    REJECT_REQUEST_EXTENSTION = 15,
    APPROVE_REQUEST_EXTENSTION = 17,
    DONE = 19,
}

export enum StatusEnumString {
    CREATED = 'CREATED',
    PENDING = 'PENDING',
    IN_PROCCESS = 'IN_PROCCESS',
    SEND_TO_CONFIRM = 'SEND_TO_CONFIRM',
    ROLLED_BACK = 'ROLLED_BACK',
    REJECTED = 'REJECTED',
    REQUEST_EXTENSTION = 'REQUEST_EXTENSTION',
    REJECT_REQUEST_EXTENSTION = 'REJECT_REQUEST_EXTENSTION',
    APPROVE_REQUEST_EXTENSTION = 'APPROVE_REQUEST_EXTENSTION',
    DONE = 'DONE',
}